// autogpt_platform/frontend/src/app/api/__generated__/models/nodeModel.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { NodeModelInputDefault } from "./nodeModelInputDefault";
import type { NodeModelMetadata } from "./nodeModelMetadata";
import type { Link } from "./link";
import type { NodeModelWebhookId } from "./nodeModelWebhookId";
import type { NodeModelWebhook } from "./nodeModelWebhook";

export interface NodeModel {
  id?: string;
  block_id: string;
  input_default?: NodeModelInputDefault;
  metadata?: NodeModelMetadata;
  input_links?: Link[];
  output_links?: Link[];
  graph_id: string;
  graph_version: number;
  webhook_id?: NodeModelWebhookId;
  webhook?: NodeModelWebhook;
}