// autogpt_platform/frontend/src/app/api/__generated__/models/link.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */

export interface Link {
  id?: string;
  source_id: string;
  sink_id: string;
  source_name: string;
  sink_name: string;
  is_static?: boolean;
}